+++ dataservices/management/commands/import_consumer_price_index_data.py
@@ -1,8 +1,9 @@
-import tablib
-from django.core.management import BaseCommand
-from import_export import resources
+import csv
+import re
 
-from dataservices.models import ConsumerPriceIndex
+from django.core.management import BaseCommand
+
+from dataservices.models import ConsumerPriceIndex, Country
 
 
 class Command(BaseCommand):
@@ -15,12 +16,28 @@ class Command(BaseCommand):
         # https://data.worldbank.org/indicator/IT.NET.USER.ZS
         # TODO refactor merge structures and more intellect import without file manipulation
         with open('dataservices/resources/Consumer_Price_Index.csv', 'r', encoding='utf-8-sig') as f:
-            data = tablib.import_set(f.read(), format='csv', headers=True)
-            internet_usage_resource = resources.modelresource_factory(model=ConsumerPriceIndex)()
-            result = internet_usage_resource.import_data(data, dry_run=True)
-            self.stdout.write(self.style.SUCCESS(result.has_errors()))
-            if not result.has_errors():
-                # No Errors lets flush table and import the data
-                ConsumerPriceIndex.objects.all().delete()
-                internet_usage_resource.import_data(data, dry_run=False)
+            file_reader = csv.DictReader(f)
+            cpi_data = []
+            for row in file_reader:
+                store = {}
+                for col_name, value in row.items():
+                    # Gather data for several years
+                    match = re.match('^\\d{4}$', col_name)
+                    if match and value:
+                        year = match.group(0)
+                        store['year'] = year
+                        store['value'] = value
+                # delete existing record so it can be updated with latest CSV data
+                # this is due to if we have updated csv in later on the year for this data
+                if row.get('ISO'):
+                    existing_record = ConsumerPriceIndex.objects.filter(country_code=row.get('ISO'), year=year)
+                    existing_record.delete()
+                try:
+                    country = Country.objects.get(iso3=row.get('ISO'))
+                except Country.DoesNotExist:
+                    country = None
+
+                country_dict = {'country_name': row.get('Country'), 'country_code': row.get('ISO'), 'country': country}
+                cpi_data.append(ConsumerPriceIndex(**country_dict, **store))
+            ConsumerPriceIndex.objects.bulk_create(cpi_data)
         self.stdout.write(self.style.SUCCESS('All done, bye!'))