+++ dataservices/management/commands/import_gdp_per_capita_data.py
@@ -1,5 +1,6 @@
 import tablib
 from django.core.management import BaseCommand
+from django.db import connection
 from import_export import resources
 
 from dataservices.models import GDPPerCapita
@@ -25,10 +26,14 @@ class Command(BaseCommand):
                 )
 
             gdp_per_capita_resource = resources.modelresource_factory(model=GDPPerCapita)()
-            result = gdp_per_capita_resource.import_data(dataset, dry_run=True)
-            self.stdout.write(self.style.SUCCESS(result.has_errors()))
-            if not result.has_errors():
-                # No Errors lets flush table and import the data
-                GDPPerCapita.objects.all().delete()
-                gdp_per_capita_resource.import_data(dataset, dry_run=False)
+            GDPPerCapita.objects.all().delete()
+            gdp_per_capita_resource.import_data(dataset, dry_run=False)
+
+            self.stdout.write('Linking countries')
+            cursor = connection.cursor()
+            cursor.execute(
+                "update dataservices_gdppercapita as d \
+                set country_id=c.id \
+                from dataservices_country c where d.country_code=c.iso3;"
+            )
         self.stdout.write(self.style.SUCCESS('All done, bye!'))