+++ dataservices/management/commands/tests/test_import_data.py
@@ -9,19 +9,22 @@ from dataservices import models
 
 @pytest.mark.django_db
 @pytest.mark.parametrize(
-    'model_name, management_cmd, object_count',
+    'model_name, management_cmd, object_count, de_rows',
     (
-        (models.CorruptionPerceptionsIndex, 'import_cpi_data', 180),
-        (models.EaseOfDoingBusiness, 'import_easeofdoingbusiness_data', 264),
-        (models.WorldEconomicOutlook, 'import_weo_data', 1552),
-        (models.InternetUsage, 'import_internet_usage_data', 264),
-        (models.ConsumerPriceIndex, 'import_consumer_price_index_data', 264),
+        (models.CorruptionPerceptionsIndex, 'import_cpi_data', 720, 4),
+        (models.EaseOfDoingBusiness, 'import_easeofdoingbusiness_data', 264, 1),
+        (models.WorldEconomicOutlook, 'import_weo_data', 1552, 0),
+        (models.InternetUsage, 'import_internet_usage_data', 264, 1),
+        (models.ConsumerPriceIndex, 'import_consumer_price_index_data', 89, 1),
+        (models.GDPPerCapita, 'import_gdp_per_capita_data', 264, 1),
     ),
 )
-def test_import_data_sets(model_name, management_cmd, object_count):
+def test_import_data_sets(model_name, management_cmd, object_count, de_rows):
     model_name.objects.create(country_name='abc', country_code='a')
+    models.Country.objects.create(name='Germ', iso2='DE', iso3='DEU')
     management.call_command(management_cmd)
     assert model_name.objects.count() == object_count
+    assert model_name.objects.filter(country__iso2='DE').count() == de_rows
 
 
 @pytest.mark.django_db
@@ -29,14 +32,13 @@ def test_import_data_sets(model_name, management_cmd, object_count):
 @pytest.mark.parametrize(
     'management_cmd',
     [
-        'import_cpi_data',
         'import_easeofdoingbusiness_data',
         'import_weo_data',
         'import_internet_usage_data',
         'import_consumer_price_index_data',
     ],
 )
-def test_import_data_sets_error(management_cmd):
+def test_error_import_data_sets_error(management_cmd):
     management.call_command(management_cmd)
     assert models.CorruptionPerceptionsIndex.objects.count() == 0
 
@@ -94,3 +96,15 @@ def test_import_comtrade_raw():
 
     management.call_command('import_comtrade_data', '--wipe')
     assert len(models.ComtradeReport.objects.all()) == 0
+
+
+@pytest.mark.django_db
+def test_import_target_age_groups():
+    management.call_command('import_countries')
+    management.call_command('import_target_age_groups')
+    data = models.PopulationData.objects.filter(country__iso1=276, year=2020)
+
+    assert len(models.PopulationData.objects.all()) == 40986
+    assert data.first().country.iso1 == '276'
+    assert len(data) == 2
+    assert data.first().age_100_plus == 4